update to go-git
Thanks for submitting #37. I decided to do the commit myself just to make sure there is nothing injected there. I updated according to your instructions and ran the go mod tidy command myself.

Sorry for not merging your commit, but there is a lot of cryptic updates there that I wanted to make sure were made with the  command.